# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorPayload": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorPayload",
        "type": "object"
      },
...
      "ApiResponse_IngestReport_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IngestReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[IngestReport]",
        "type": "object"
      },
...
      "ApiResponse_KnowledgeDocumentDTO_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeDocumentDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[KnowledgeDocumentDTO]",
        "type": "object"
      },
      "ApiResponse_KnowledgeDocumentListData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeDocumentListData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[KnowledgeDocumentListData]",
        "type": "object"
      },
...
      "ApiResponse_ReindexReport_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReindexReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[ReindexReport]",
        "type": "object"
      },
...
      "ApiResponse_dict_str__object__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[dict[str, object]]",
        "type": "object"
      },
      "ApiStatus": {
        "enum": [
          "success",
          "accepted",
          "error"
        ],
        "title": "ApiStatus",
        "type": "string"
      },
...
      "Body_ingestKnowledgeDocument": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "metadata_json": {
            "minLength": 2,
            "title": "Metadata Json",
            "type": "string"
          }
        },
        "required": [
          "file",
          "metadata_json"
        ],
        "title": "Body_ingestKnowledgeDocument",
        "type": "object"
      },
...
      "IngestReport": {
        "additionalProperties": false,
        "properties": {
          "chunks_created": {
            "minimum": 0.0,
            "title": "Chunks Created",
            "type": "integer"
          },
          "chunks_skipped": {
            "minimum": 0.0,
            "title": "Chunks Skipped",
            "type": "integer"
          },
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "index_version": {
            "title": "Index Version",
            "type": "string"
          },
          "pages_extracted": {
            "minimum": 0.0,
            "title": "Pages Extracted",
            "type": "integer"
          },
          "pages_total": {
            "minimum": 0.0,
            "title": "Pages Total",
            "type": "integer"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "doc_id",
          "sha256",
          "pages_total",
          "pages_extracted",
          "chunks_created",
          "chunks_skipped",
          "index_version"
        ],
        "title": "IngestReport",
        "type": "object"
      },
...
      "KnowledgeDocumentDTO": {
        "additionalProperties": false,
        "properties": {
          "allowed_for_demo": {
            "title": "Allowed For Demo",
            "type": "boolean"
          },
          "citation_text": {
            "title": "Citation Text",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "license_note": {
            "title": "License Note",
            "type": "string"
          },
          "material_aliases": {
            "items": {
              "type": "string"
            },
            "title": "Material Aliases",
            "type": "array"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/KnowledgeSourceType"
          },
          "status": {
            "$ref": "#/components/schemas/KnowledgeDocumentStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "doc_id",
          "title",
          "source_type",
          "sha256",
          "citation_text",
          "status",
          "license_note",
          "allowed_for_demo",
          "created_at"
        ],
        "title": "KnowledgeDocumentDTO",
        "type": "object"
      },
      "KnowledgeDocumentListData": {
        "additionalProperties": false,
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeDocumentDTO"
            },
            "title": "Documents",
            "type": "array"
          }
        },
        "title": "KnowledgeDocumentListData",
        "type": "object"
      },
      "KnowledgeDocumentStatus": {
        "enum": [
          "ready",
          "indexing",
          "unavailable",
          "disabled"
        ],
        "title": "KnowledgeDocumentStatus",
        "type": "string"
      },
      "KnowledgeSourceType": {
        "enum": [
          "paper",
          "report",
          "material_note",
          "other"
        ],
        "title": "KnowledgeSourceType",
        "type": "string"
      },
...
      "ReindexReport": {
        "additionalProperties": false,
        "properties": {
          "chunks_indexed": {
            "minimum": 0.0,
            "title": "Chunks Indexed",
            "type": "integer"
          },
          "chunks_skipped": {
            "minimum": 0.0,
            "title": "Chunks Skipped",
            "type": "integer"
          },
          "documents_indexed": {
            "minimum": 0.0,
            "title": "Documents Indexed",
            "type": "integer"
          },
          "index_version": {
            "title": "Index Version",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "documents_indexed",
          "chunks_indexed",
          "chunks_skipped",
          "index_version"
        ],
        "title": "ReindexReport",
        "type": "object"
      },
      "ReindexRequest": {
        "additionalProperties": false,
        "properties": {
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          }
        },
        "title": "ReindexRequest",
        "type": "object"
      },
...
      "UpdateKnowledgeDocumentRequest": {
        "additionalProperties": false,
        "description": "Enable or disable one already-indexed knowledge document.",
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "UpdateKnowledgeDocumentRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Traceable SEM nanoparticle analysis and evidence-backed query API",
    "title": "NanoLoop Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/documents": {
      "get": {
        "operationId": "listKnowledgeDocuments",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeDocumentListData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "List Documents",
        "tags": [
          "knowledge"
        ]
      },
      "post": {
        "operationId": "ingestKnowledgeDocument",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingestKnowledgeDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_IngestReport_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "Ingest Document",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/documents/{doc_id}": {
      "patch": {
        "operationId": "updateKnowledgeDocument",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKnowledgeDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeDocumentDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "Update Document",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/reindex": {
      "post": {
        "operationId": "reindexKnowledge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReindexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReindexReport_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "Reindex Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
...
  }
}